requirements.txt
streamlit
langchain
requests
python-dotenv
streamlit-chat
pandas
langchain-google-genai